# Optimize key formatting