An optimization to prevent allocating a new string for each key in the json response. Prior to this change,
```ruby
json.set! :foo, 'bar'
```

would allocate a string for `foo` on every render. By using `Symbol#name`, it will instead use the frozen interned string which saves on that extra memory allocation.

```ruby
my_symbol = :foo

Benchmark.memory do |x|
  x.report('Symbol#name') { 100.times { my_symbol.name } }
  x.report('Symbol#to_s') { 100.times { my_symbol.to_s } }

  x.compare!
end
```

```
Calculating -------------------------------------
         Symbol#name     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
         Symbol#to_s     4.000k memsize (     0.000  retained)
                       100.000  objects (     0.000  retained)
                         1.000  strings (     0.000  retained)

Comparison:
         Symbol#name:          0 allocated
         Symbol#to_s:       4000 allocated - Infx more
```

Using `name` is also more performant

```
ruby 3.4.4 (2025-05-14 revision a38531fd3f) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
         Symbol#name     2.300M i/100ms
         Symbol#to_s     1.318M i/100ms
Calculating -------------------------------------
         Symbol#name     40.122M (± 1.3%) i/s   (24.92 ns/i) -    202.358M in   5.044457s
         Symbol#to_s     17.681M (± 2.3%) i/s   (56.56 ns/i) -     89.620M in   5.071414s

Comparison:
         Symbol#name: 40121894.9 i/s
         Symbol#to_s: 17681248.1 i/s - 2.27x  slower
```